chore: Rename internal/db/{migrations->migration/schemas}
