+++ internal/database/dbconn/connect.go
@@ -6,7 +6,7 @@ import (
 	"github.com/hashicorp/go-multierror"
 	"github.com/prometheus/client_golang/prometheus"
 
-	"github.com/sourcegraph/sourcegraph/internal/database/migrations"
+	"github.com/sourcegraph/sourcegraph/internal/database/migration/schemas"
 )
 
 // ConnectRawForTestDatabase connects to the given data source and returns the handle. After successful
@@ -16,7 +16,7 @@ import (
 // This function returns a basestore-style callback that closes the database. This should be called instead
 // of calling Close directly on the database handle as it also handles closing migration objects associated
 // with the handle.
-func ConnectRawForTestDatabase(dsn string, schemas ...*migrations.Schema) (*sql.DB, func(err error) error, error) {
+func ConnectRawForTestDatabase(dsn string, schemas ...*schemas.Schema) (*sql.DB, func(err error) error, error) {
 	return connect(dsn, "", "", schemas)
 }
 
@@ -37,7 +37,7 @@ func ConnectRawForTestDatabase(dsn string, schemas ...*migrations.Schema) (*sql.
 //
 // Note: github.com/jackc/pgx parses the environment as well. This function will also use the value
 // of PGDATASOURCE if supplied and dataSource is the empty string.
-func connect(dsn, appName, dbName string, schemas []*migrations.Schema) (*sql.DB, func(err error) error, error) {
+func connect(dsn, appName, dbName string, schemas []*schemas.Schema) (*sql.DB, func(err error) error, error) {
 	cfg, err := buildConfig(dsn, appName)
 	if err != nil {
 		return nil, nil, err
