+++ internal/database/dbtest/dbtest.go
@@ -15,7 +15,7 @@ import (
 	"github.com/lib/pq"
 
 	"github.com/sourcegraph/sourcegraph/internal/database/dbconn"
-	"github.com/sourcegraph/sourcegraph/internal/database/migrations"
+	"github.com/sourcegraph/sourcegraph/internal/database/migration/schemas"
 )
 
 // NewTx opens a transaction off of the given database, returning that
@@ -145,9 +145,9 @@ func initTemplateDB(t testing.TB, config *url.URL) {
 
 		cfgCopy := *config
 		cfgCopy.Path = "/" + templateName
-		_, close := dbConnInternal(t, &cfgCopy, []*migrations.Schema{
-			migrations.Frontend,
-			migrations.CodeIntel,
+		_, close := dbConnInternal(t, &cfgCopy, []*schemas.Schema{
+			schemas.Frontend,
+			schemas.CodeIntel,
 		})
 		close(nil)
 	})
@@ -174,7 +174,7 @@ func dbConn(t testing.TB, cfg *url.URL) *sql.DB {
 	return db
 }
 
-func dbConnInternal(t testing.TB, cfg *url.URL, schemas []*migrations.Schema) (*sql.DB, func(err error) error) {
+func dbConnInternal(t testing.TB, cfg *url.URL, schemas []*schemas.Schema) (*sql.DB, func(err error) error) {
 	t.Helper()
 	db, close, err := dbconn.ConnectRawForTestDatabase(cfg.String(), schemas...)
 	if err != nil {
