+++ internal/database/migration/schemas/schema.go
@@ -1,4 +1,4 @@
-package migrations
+package schemas
 
 import (
 	"io/fs"
